admission: allow skipping setting the home env
This allows using whatever the image comes with, usually useful for
buildpack based images.

Also removes any workingDir if it's set when NO_HOME is passed so the
image will specify what to do.

Some refactor to the tests too, adding new ones but not changing the
essence of the existing ones.

Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: If4509c33d5dcc10a743a88c5aa6a3699609d2951